requirements-test.txt
coverage>=4.0.2,<5
mock>=1.3.0,<4
tox>=2.2.1,<3
django-taggit>=0.21.3
fake-factory>=0.4.0
factory-boy>=2.4.1

